pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "griffon"
version = "0.6.5"
description = "A tool to exploit bad configurations in Active Directory"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "GPL-3.0"}
authors = [
    {name = "GriffonAD Contributors"}
]
dependencies = [
    "ldap3",
    "dash",
    "dash-cytoscape",
    "jinja2",
    "colorama",
    "pycryptodome",
]

[project.scripts]
griffon = "griffonad.griffon:main"

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["griffonad*"]

[tool.setuptools.package-data]
"griffonad" = ["config.ml"]
"griffonad.assets" = ["*"]
"griffonad.tools" = ["comment.cmtx", "pol.json"]
